update: contato/mensagem

+++ src/main/java/com/bedinviagens/repository/PacoteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bedinviagens.model.Pacote;
 
-public interface PacoteRepository extends JpaRepository<Pacote, Integer> {
+public interface PacoteRepository extends JpaRepository<Pacote, Long> {
 
 }
